Merged PR 252801: Git MVC Quickstart review
This PR is to review the Git quickstarts for the migration - most content is the same as before, except for updated MVC styling, and new artifcts like the Visual Studio/Command Line tabs. The major change here is I also imported the IntelliJ content from Content.Java into our repo per Jeremy, and added a few steps at the top so it was self contained.

Staged link should be live soon
https://review.docs.microsoft.com/en-us/vsts/git/create-repo-intellij?branch=users/sdanie/git_mvc_quickstarts