Merged PR 257420: Made some text more clear.
When I re-read the following text:

**Please refer to the SharePoint topic: [Upgrade to SharePoint Server 2016], for full instructions on upgrading your SharePoint server.** This topic provides additional guidance to ensure the TFS SharePoint sites are functional after upgrading to SharePoint 2016.

The words "This topic" in the second sentence felt like we were referring to the SharePoint topic. So I changed it to "The instructions below", to make it more clear.
I know we aren't supposed to say "below", but couldn't think of a better way to say it.