Merged PR 328646: Document that "Copy Files Over SSH" is intended for Linux target machines.
A GitHub issue came up complaining that the task did not work correctly when trying to copy files to a Windows machine running Win32-SSH.  The problem is that some command lines baked into the task assume that the remote shell will be a Linux machine.  https://github.com/Microsoft/vsts-tasks/issues/6826

We need to clarify in our documentation that the task is meant for Linux.