Add info about the Download Secure File task
This page seems to act as a general overview of Secure Files, but does not provide any information about how to consume the secure files outside of integrating the functionality in a custom task. I've included a link to the Download Secure File Utility task in order to provide this information.