Merged PR 343681: Third-party extension developer guidance for support public projects
Guidance for extension developers that want to make their extension available to non-member users (e.g. anonymous users) in the context of a public project.

[docs review link](https://review.docs.microsoft.com/en-us/vsts/extend/develop/public-project?view=vsts&branch=features%2Fthird-party-extension-public-projects)

Note: it is OK/expected for these docs to go live prior to the public preview of public projects, as the feature has already been announced and we want extension developers to start changing their extensions ahead of broader public availability.