Merged PR 373484: edit pass: faq-add-team-members
Please review the copy edit to this article and let me know if I’ve changed the technical meaning anywhere. This is for VSTS work request 1302753.

Comments:

Line 69: **Can Azure AD B2B users be added as members instead of guests?**
Is this an article or a section that should have a hyperlink?

Thanks,
Trisha

Related work items: #1302753