Merged PR 376919: Added NuGet task to TOC per gh1559
https://github.com/MicrosoftDocs/vsts-docs/issues/1559

- Added link to index page [here](https://review.docs.microsoft.com/en-us/vsts/pipelines/tasks/?view=vsts&branch=pr-en-us-376919#tool)
- Added NuGet tool installer to TOC, view that page [here](https://review.docs.microsoft.com/en-us/vsts/pipelines/tasks/tool/nuget?view=vsts&branch=pr-en-us-376919) to see it on the TOC

If I didn't know the link to that index page (from the GI issue) to paste it into the browser, I wouldn't have found it. It doesn't seem to be linked from anywhere. I tried to find it by browsing from the main Azure DevOps index, down to the pipelines index, but then could not find it.