Merged PR 382421: JavaScript pipeline updates
Taking Vijay's guide and adding in items from the CDA conversation, and correcting/enhancing other tasks to better fit where the ecosystem's at.

Work in progress and I have additional items and I'm adding , but need to see what it's like in preview due to my unfamiliarity with the moniker/tab minefield going on in this doc.