Images not linked correctly
I'm not 100% sure this is correct, but it's definitely not worse than it was. :) Check with the images referenced at the top of the page to make sure I got the png paths correct. Seems like the top image uses a slightly different relative path and I'm not sure where these images actually exist.