Merged PR 388489: Updated status of default fork builds on GitHub and GHE
Fork builds will be enabled by default for both GitHub and GitHub enterprise. Passing of secrets will also be enabled by default for GitHub Enterprise.
See !388483 and !388086 as well as the related work items.

Feedback on how to phrase this and what info should be added is welcome.

Related work items: #1343212