Merged PR 389481: edit pass: project-admin-tutorial
Please review the copy edit to this article and let me know if I’ve changed the technical meaning anywhere. This is for Azure DevOps work item 1337708.

Comments:

There are several places where text seems to be duplicated. I’ve noted it here but didn’t delete it in case it’s needed for markup or other reasons.
Line 30: This sentence sounds awkward to me, but I’m not sure how to rewrite it. Can you check it and revise if needed? “Larger organizations will want to consider using Azure Active Directory to keep the maintenance of managing permissions and user access.”
Line 39: Can you check the article link? It didn’t work for me.
Line 70: This seems to duplicate some of line 65. Can you check and delete if it’s not needed?
Line 104 to 107: These lines seem to duplicate lines 97 to 100. Can you check and delete if they’re not needed?
Lines 135 to 136: These lines seem to duplicate lines 122 and 123.  Can you check and delete if they’re not needed?
Line 213: This line seems to duplicate line 201. Can you check and delete if it’s not needed?
Line 216: Does this link work? (/tfs/server/index)
Lines 214 to 216: These lines seem to duplicate lines 207 and 209. Can you check and delete if they’re not needed?

Related work items: #1337708