Merged PR 390163: edit pass: devops-alm-overview
Please review the copy edit to this article and let me know if I’ve changed the technical meaning anywhere. This is for Azure DevOps work item 1337698.

Comment:
The images in line 61 don't show in VS Code or in the repo. The images in line 69 are duplicates in the middle and on the right.