Merged PR 389018: edit pass: provide-feedback
Please review the copy edit to this article and let me know if I’ve changed the technical meaning anywhere. This is for Azure DevOps work item 1337709.

The alt text "TFS About page, version number" is showing in the preview and might not render properly in the published article.

Related work items: #1337709