Update documentation on auto login and lock
Update documentation on auto-login and lock to contain a second link which has the original author's plist.  The author was hosting the file but their links appear to be broken.  Including this would have saved me some time.

If you want to chat privately about this, feel free to reach out to my MSFT alias: jiboulte.