Merged PR 405232: Manual sync public changes to private
This PR brings in customer (and some internal) PRs that were reviewed, approved, and merged into the public repo that didn't get synced back to the private repo. See the commits list for the PRs, unfortunately there doesn't seem to be a way to get from the files in the files view to the associated commit and commit message (which is the link to the PR). All changes in this current PR have been reviewed and approved over on the public repo - this review is just a sanity check.