Merged PR 395799: DO NOT MERGE - Analytics support for Azure DevOps Server 2019 RC1
A comprehensive set of updates to the Analytics docs, to reflect we support it for RC1 and beyond.
Because this branch contains moniker configuration changes, it shouldn't be merged into master until Nov 19th and coordinated with updates to the master branch.

This will be the major pull request for all changes. All reviewers must approve before approving.

Related work items: #1281711, #1377991, #1377992