Update transforms-variable-substitution.md
Added sentence the re-enforces that the parameters.xml file should be used to handle updating values that are not within the `applicationSettings`, `appSettings`, `connectionStrings`, and `configSections` elements. Added information that indicates that variables are not updated within parameters.xml file, and has to be handled using 3rd party task such as "Replace Tokens".