I fixed a few spelling and grammatical errors.
Fixed a few spelling errors. Fixed a couple of what I thought were grammatical errors.  There are also a couple of places where I felt an oxford comma would be better, but depending on the style guide you subscribe to, it is not required, so I did NOT mark those.