Updating URL samples to new format in npmrc.md
The previous URLs used in the Create a token that lasts longer than 90 days section were using the old format of URL for the artifacts NPM feeds, not the new style that are now being used. This caused me to waste a long time debugging an authentication error as I didn't notice the difference in URL format.