Merged PR 473619: Troubleshooting authentication issues for Azure Pipelines app
I refactored these procedures into a set of include files (one for teams and one for slack) that then used other include files, so that there is no duplication. Then I added a Troubleshooting step at the bottom of each article. Here are the staged versions:

* [Teams - troubleshooting](https://review.docs.microsoft.com/en-us/azure/devops/pipelines/integrations/microsoft-teams?view=azure-devops&branch=pr-en-us-473619#troubleshooting)
* [Slack - troubleshooting](https://review.docs.microsoft.com/en-us/azure/devops/pipelines/integrations/slack?view=azure-devops&branch=pr-en-us-473619#troubleshooting)

From my side this PR is ready to go, so if it looks good to you, please complete the PR and I will ensure it gets published. If I need to make any changes, let me know in the PR and I'll fix it first thing tomorrow morning.