Update xamarin.md
We're proposing this change to the docs because we recently had a hard time configuring the Xamarin.iOS build task to produce an .ipa package for CI/CD in Azure Pipelines. We had to troubleshoot a lot and we even had to perform the build using a regular bash task to finally find out that we were passing a wrong apple certificate. The Xamarin.iOS task never through an error but we noticed no .ipa file was being generated at all by carefully reviewing the folders of the agent, somehow a regular bash msbuild command provided more information pointing we were missing signing identities for the app bundle id.

After understanding the issue we guessed it be good idea to share this with our customer through this docs.
Hopefully, you'll find our contribution useful. Proposing this with @ManuelGalindo we're both Azure DevOps support engineers.