Update azure-resource-group-deployment.md
It is very unclear in the documentation what value should be included here. Especially since the value to be provide should be the GUID/resourceID of the service connection created for the Azure subscription being used. I propose to add a little more text to that particular field to indicate that.