Merged PR 497124: Update docs to show new "Set up the new Azure Artifacts Credential Provider"
The artifacts "Connect to Feed" dialog now shows a new option to install the Azure Artifacts Credential Provider, and no longer advertises the older credentialProviderBundle.zip.  This PR updates the screenshot to show the modified UI.

The new credential provider also requires nuget 4.8+, so the minimum recommended nuget version is also updated (was 3.5+).