Updated App Type
Updated App Type to be required, and say that it's the function app type not web app type. Since the task.json file shows that it's a required type and actually referring to the runtime of the function app. https://github.com/microsoft/azure-pipelines-tasks/blob/master/Tasks/AzureFunctionAppV1/task.json