Uncomment incorrectly commented-out line of code
In the code snippet describing the code required to launch an Android Emulator on the Azure Pipeline agent, the line which invokes `sdkmanager` to download the emulator image is commented out. However, this emulator image is not preloaded on the agent VMs, so this line must be uncommented for these instructions to work.