Resolve syncing conflicts from repo_sync_working_branch to master (#194)
* Update AppCenterDistributeV3.md

"destinationType == store" instead of "Store" as uppercase is not supported

* Fixed casing of parameters

The parameters for *deploymentMode* should all start with capital letters.

* Update download-secure-file.md

Adding mention that files are downloaded to `$(Agent.TempDirectory)`. We have seen this information needed for customers who require uploading files in bulk or using tasks that require a directory and a file name, rather than an absolute path, such as the FTP task.

* added YAML example for iot edge deployment

* Publishing to live 10/7/2019 (#186)

* Updated App Type

Updated App Type to be required, and say that it's the function app type not web app type. Since the task.json file shows that it's a required type and actually referring to the runtime of the function app. https://github.com/microsoft/azure-pipelines-tasks/blob/master/Tasks/AzureFunctionAppV1/task.json

* codecoverage for pull requests

* Azure Artifacts landing page converted

* Converted Reporting landing page

* fixed errors, warnings, and most suggestions

* more fixes

* another build fix

* finishing up for the day, final tweaks

* removed index landing pages from TOCs.

* Added images and updated doc

* Update per feedback, AB#1589636, AB#1589632

* Fixed build syntax error

* Review tweaks

* Review tweaks

* Final review tweaks

* Task index updates

* Updated task index

* Task list updates

* Updated task list

* Intro tweak

* addressing vsts-docs issue 5830

* Added icons

* Icons

* More icons

* Icons

* Icons

* Tasks

* Tasks

* More icons

* Restored file

* SonarQube

* SonarQube

* Final icons

* Icons

* Refresh python-web-app.md (#153)

* first pass

* corrections

* fix graphics etc

* changes

* few more changes

* some updates

* few fixes

* changes per peer review

* touchups

* adds conceptual versioning for custom upstream src (#185)