Added old flag back for reference in ignore-file.md
The documentation on --assume-unchanged was removed which caused problems for my team who looked at the previous version that only had that flag and not the new preferred one of --skip-worktree. In light of the confusion it makes sense to include both in the documentation for anyone who used the old flag previously (since they need to know how to resume change tracking).