provide yaml keywords for the built in conditions
I derived these yaml keywords (is that what you'd call them?) by using the view yaml UI in a build definition.  If this is accurate, it'd be great if this made it into the Azure Pipelines YAML vs code extension's Intellisense.  I don't know how to do that which is why I didn't, if you do, please reference the PR here so I can learn.