List of branches to trigger pipelines doesn't work
Documentation give the impression that you can use a plain list of branches to define the ones that will trigger the pipeline, but that doesn't works. It needs to use the "long" version, specifying the `branches` and `includes` nodes. This is similar to how it's defined the default value some lines under.