Fix incorrect default value
If a branch is not specified for a pipeline resource, all branches are included in artifact selection, not just the master branch. This is stated in the YAML schema page (https://docs.microsoft.com/en-us/azure/devops/pipelines/yaml-schema?view=azure-devops&tabs=schema#resources) and appears to be true from my own testing.