Remove outdated note at end of page
This note section links to https://github.com/microsoft/azure-pipelines-task-lib/blob/master/node/docs/stepbystep.md, which has been cleared, and now simply links right back to this doc page. It seems the docs were updated to use the Typescript API, but this old link was missed in the process.