Add Build.SourceBranch since it is also available within a template expression
In testing, I found that I was able to access `Build.SourceBranch` in a template expression by wrapping the variable using this syntax:

`${{ if contains(variables['Build.SourceBranch'],'vNext/') }}` 

I did this in a fit of desperation to try to get a value used by a demand.  See issues #7494 , #7170, #5279 for additional context.