Update condition docs to cover cancelation
Call out that the succeeded function also
returns false when the pipeline is canceled.
Update the example using a variable in a condition
to reflect the best practice of combining it with
the succeeded function. The example as previously
written encouraged a pattern that leads to stages/jobs
executing even when the pipeline is canceled, which
is not usually what people want.