Completion of the lab is first dependent on creating a Service Connection of type ARM
To ensure your Azure DevOps project has the needed authorization through to your Azure Subscription, create a Service Connection of type ARM.  This will be required when creating a pipeline within the project to deploy to Azure Kubernetes Service.  Otherwise the dropdown lists for Cluster and Container Registry are empty.  Please insert detailed instructions or link to an appropriate doc to avoid our customers bumping into the same.