Update scheduled-triggers.md
I'm adding this "Tip" to the documentation as one of our devs ran into this issue when using days of the week.  The UI has checkboxes for Mon - Sun and the example above has Mon-Fri.  Following this logic, we used Mon-Sun in the cron task itself, but this then creates a schedule of only two days, Sun and Mon.  This did not seem, per se, to be an actual bug with the code itself, but more of a need to understand how it was generated.
Using `0 5 * * Mon-Sun` generates scheduled runs of midnight EST on Sun and Mon.
Using `0 5 * * Sun-Sat` generates scheduled runs of midnight EST every day

I tried to make the tip a little bit convoluted to show the inclusion of Sunday while skipping certain days.