Getting ref of non-self repo causes exception
When you run following example you will get exception on validating pipeline

> An error occurred while loading the YAML build pipeline. Object reference not set to an instance of an object

I was trying to figure out if this is a bug in documentation or Azure DevOps itself but based [on this](https://developercommunity.visualstudio.com/content/problem/1252381/object-reference-not-set-to-an-instance-of-an-obje-62.html) I had to assume that this is doc issue.

I was trying to find out what is in `resources.repositories` but I couldn't make it. Probably whole `Repository details` should be updated or deleted as it doesn't provide any value at the moment.