Update azure-file-copy.md
The table of arguments and description of arguments did not match the actual inputs to the yaml task. For example, in the yaml one would type storage: {name of existing storage account}, but the table had the argument as RM Storage Type, which can create confusion for beginners as that is not a listed input in the yaml snippet. I changed all the Arguments in the Table to match the actual input for the yaml task.