When the users are removed from organization
My customer wants to know if the users are removed from the organization, will the work items assigned to the users be affected.
Based on my test, the work items will not be removed, and the settings in them will not be changed as well. Could you kindly update something about this issue in the document?