Update set-your-preferences.md
I saw this article while working on related content that references this one, and I spotted these items. The "from" corrections are for style and consistency. The "IE" correction is because we no longer use that acronym. (By contrast, the "UI" acronym is appropriate — especially for this audience — because that's still in use.) Thanks!