Update extract-files.md
Proposing an update for the description of the overwriteExistingFiles-option. The motivation behind the proposed change is to try to describe more explicitly how the overwriteExistingFiles option works in the case of existing files. One might assume that when the overwriteExistingFiles option is set to false no user interaction is needed when there is existing files in the output directory.