Update service-endpoints.md
Changed below line #176:

The user who created the service connection is automatically added to the project-level Administrator role for that service connection. Users and groups assigned the Administrator role at the hub-level are inherited if the inheritance is turned on.

To: below (without "project-level" as the user is only added as "Administrator" on the Service Connection he creates. He is not added as "Administrator" at the Project-level security:

The user who created the service connection is automatically added to the Administrator role for that service connection. Users and groups assigned the Administrator role at the hub-level are inherited if the inheritance is turned on.