# Set Variables in Scripts - Better readability and more fleshed-out examples around setting variables in scripts. - More direct links to relevant sections of related documentation.
# Logging Commands
- Provided an excerpt from "Set Variables in Scripts" that indicates, local to the function documentation, that there are different syntaxes for output variables depending on their relative consumption location.

I burned a lot of time not realizing that output variables in the same job with `isoutput=true` were access via `taskName.varName`. This was exacerbated by the link text being  "set variables in scripts" while actually linking  to a similarly named part of "Define variables" when "Set variables in scripts" is actually the relevant page.