Improve convertToJson example
Using arbitrary text as part of the _source code_ of a program is
error-prone and unsecure, so this change replaces the implicit
suggestion to do so with a simpler, safer, more robust solution.

👉 https://github.com/MicrosoftDocs/azure-devops-docs/issues/11983#issuecomment-1125092780

`# Example output` is not valid JSON, so this change also removes it
from the `json` code block.

Resolves #11983.