Squash PR (part 2)
Squash merge commit doesn't have reference to source branch, it is absolutely new commit with all changes. Furthermore it has a warning about recommendation of deleting source branch to prevent confusion with it.

PR to source branch back will suppose this commit as different one. So all changes of squash merge commit will be displayed in this PR.