Validating and troubleshooting webhooks
I'm adding a validating and troubleshooting section to the webhook documentation as I have spent multiple hours trying to figure out my webhook issue.  Each time it has been delivering my pipeline YAML to a feature branch (not the default) and creating the pipeline, pointing my pipeline to my feature branch YAML, and clicking 'RUN'.  This *looks* to all effects that it should work, but doing this does not seem to register the webhook to my organization's distributed tasks, thus giving me a 500 status code response every time that the webhook doesn't exist.  The steps above allow the user to deliver their code to a feature branch, and save and queue it from there.  From this point on, the webhook is registered correctly, and responds with a 200 status code.  The means of *getting* to the above via the UI are also not very apparent, so hopefully this will assist others who are also struggling with this.  I had only found 1 or 2 references when trying to research the problem I was facing.