Update to the steps
Hi, I just submitted part of this document's update a couple of days ago, but one important section was removed before the approved copy above.  I've added that part back in again.  The "more actions" menu differs between the pipeline yaml page for the "Edit" screen and after opening the pipeline.  If the "Edit" button is not clicked first, the user cannot get to the "hidden" menu options for the pipeline.  I don't know if this is actually more of a UI issue that should be fixed by another team, but until then, this is the way to get to this hidden menu :) Thanks.