Update FAQ in pipeline-artifacts.md
The opening sentence of this document says:

> You can also publish your artifact to a file share...

But towards the end, in the FAQ section we have this:

> **Q: Do you support publishing artifacts to a shared folder?**
> A: Not currently, but this feature is planned.

It appears that publishing to a shared folder was implemented at some point but the FAQ was not updated. I have now removed the question because the docs say directly that publishing to a shared folder is supported. This change is valid assuming that "file share" and "shared folder" refer to the same thing (which I think they do!).