Update multi-repo-checkout.md
On Line 272, removed this part of the line ", unless you explicitly override the behavior during checkout." Because the triggering repository is not automatically checkout (an internal bug was created and this information is tested and verified with PM).   

In the Yaml, added lines 297 to 300, so that the information in the Table that follows will still be applicable.