Fix bash snippet for cloning with a PAT
The example provided is totally wrong. Used as is, it doesn't work.

The main problem is that the comment tells the user to prefix their PAT with a 
`:`, but then, on the next line, there is also a `:` added in the `printf`. 
This ends up generating a wrong base64.

The printf is also completely wrong. No space between `"%s"` and `":$MY_PAT"`, 
means it's just one string and no 2nd argument for the `"%s"`, making the 
placeholder unused. The proper form is to have the `:` in the format string and 
passing "$MY_PAT" as an argument.