Update: added steps, muted keys/id File: auditing-streaming.md
Update 1: Added additional step 3 line 114 to include the expansion of a tab in azure Log Analytics Workspace to reflect the new view in Azure portal. 

Update 2: Muted/hid the workspace ID and keys.

ACTION REQUIRED: I need to update the .png file that is on line 117 and replace with new one.