Update change-organization-location.md
As of Aug 25, 2023, we'll be showing the orgs geography, in addition to the region on the org settings page. At that point, we'll want to make this set of changes to shift the focus of this article from region to geography. There are still a few references to region that need to remain.

NOTE: Please check on two things:
1) I didn't update the screenshot with this PR, the one showing org settings (step 3) should be updated once this change is in the FabrikamFiber ring.
2) I wasn't sure how the "In this article" content is being generated, but we may need to change those two references from region to geography (unless they're generated based on the headers I already changed - just wanted to double check)