Update VMSS creation docs to explicitly use Uniform orchestration
The default orchestration mode for VMSS in the Azure CLI module will change to Flexible soon. Reference: https://techcommunity.microsoft.com/t5/azure-compute-blog/breaking-change-for-vmss-powershell-cli-customers/ba-p/3818295

Updating our docs so that the orchestration mode is explicitly set to Uniform.
This is to avoid scenarios in which customers follow the docs and fail because their organisation doesn't support flexible orchestration VMSS pools.