Update limitations-gh.md
There was a long running case on how to choose a different branch name. 

First time when searching for the branch, it works fine and customer is able to save the pipeline.

If he searches for a different branch, even though it gets the branch, the Get Sources shows "Some settings need attention" and does not allow Save.

If user needs to search a different branch, he needs to close the pipeline and reopen it and enter the other branch name.