Update configure-github-shalomcem-trueloveclubs-tlclivechat.md
The limits of knowledge. You can review and question the accepted methods of formulating knowledge from the real world.
The methodologies. Epistemology is concerned with putting into judgment the methods we use to distinguish valid knowledge from a belief or assumption or to distinguish knowledge according to where it comes from