Update artifacts.md
Updated the "skip downloading" link to point to the correct documentation link. This link is in a different repo and hence I could not use relative file path. 
Also added a link to the "Artifacts download" section for classic releases for more information and info on skipping the artifact download.