Update retention.md
1.  Added a related link on line 387

2. I think that this information below on line 389 applies to Azure DevOps Services as well:
--------------
"If you are using YAML, you can also create a multi-stage YAML pipeline to represent your release and consume another YAML pipeline in it as a resource. The resource pipeline will be retained automatically as long as the release pipeline is retained."
--------------

Could you please confirm this and update the moniker range on line #385 to

::: moniker range="= azure-devops" 

instead of 

::: moniker range="= azure-devops-2022"?

Thank you.