Update publish.md
Adding an example for the nuget sources command without a nuget.config file as that is what I eventually needed in order for my nuget package upload process to work within a GHEC Actions pipeline execution. I needed to run the nuget sources command whenever the pipeline executed, and providing the PAT as a Github Actions secret, and referencing the PAT secret "variable" accordingly. Without nuget.config file examples, I was unaware that the command could simply omit the config file reference, and I wanted to contribute that example here.