Update migrate-integrate.md
Your comment of giving link to VSTS marketplace's homepage is duly noted. However, we see other links present on this page going to their respective marketplace listing and another one to the website's link. 

Let me know if this PR commit is fine or not in the light of above observations. 

Thank you for all the support! :)