Update the Exclusive Lock Behavior code examples.md
The Exclusive lock behavior pipeline examples are not using the protected resource (for example Environment) on which the Exclusive lock is enabled. 
So, the pipeline runs would not get a lock to the protected resource and will not honor the lockbehavior.

Updated the code samples to use the environment on which the Exclusive lock is enabled.