Update dashboards.md to match screenshot
I noticed that the screenshot had a button labeled "Create", not "Save". This matches with the current UI in my Azure DevOps instance. I also noticed that the Add a Widget page doesn't automatically open when I select Create, but I am reluctant to edit the other steps in case that's due to something odd on my end...