attempting to fix unclosed anchor tag
On the rendered page, in the table under "Fields associated with links and attachments", everything in between "External Link Count" and "Remote Link Count" is treated as if it were a giant link. The link doesn't close until it hits some Markdown link syntax within the second column on the row, "Remote Link Count". This seems to be caused by the anchor tag near "External Link Count" but it's kind of hard to tell without seeing a preview build.