Update billing-faq.yml to include warning for removing billing
The FAQ items for the Azure portal integration instruct customers to remove billing and reconfigure if the Azure resource details are incorrect. I added a warning to these two questions to ensure customers know that doing these steps will temporarily revert the org to the free tier to avoid surprise disruptions to their users.